magit-list-refs: support a format argument
Extend magit-list-refs to take a format argument to allow the caller
to specify an arbitrary format, and then update magit-list-refnames to
use magit-list-refs underneath.  These two functions are currently
simple enough that this second step doesn't gain much in terms of
reducing duplication.  However, the next commit will extend
magit-list-refs to support a sorting argument, so having
magit-list-refnames build on magit-list-refs prevents us from having
to do that in two places.

Also, rename the "args" argument of magit-list-refnames to
"namespaces", and restrict its callers to passing a string or a list
for this argument rather than supporting a &rest argument list.  This
change isn't necessary, but it results in consistent signatures
between the two functions.

Now that magit-list-refs takes a format argument, we should consider
removing magit-format-refs, which supports an arbitrary format
specification as well as some additional field-splitting logic.  It
was added added in 3c4c28f4 (magit-format-refs: new function,
2015-01-27) but has yet to be used anywhere.